+++ docs/.vitepress/config.ts
@@ -1,13 +1,18 @@
 export default {
   base: "/qalendar/",
   title: "Qalendar",
-  description: "Vue 3 calendar component",
   themeConfig: {
     nav: [
-      { text: "Home", link: "/" },
+      { text: "Guide", link: "/guide" },
       { text: "v1 to v2", link: "/migration" },
       { text: "Roadmap", link: "https://github.com/tomosterlund/qalendar/projects/1" },
       { text: "Github", link: "https://github.com/tomosterlund/qalendar" },
     ],
+    footer: {
+      message: 'Released under the MIT License.',
+      copyright: 'Copyright © 2022-present Tom Österlund'
+    },
+    logo: 'https://discover-test-files.s3.eu-central-1.amazonaws.com/Q+MAIN.png',
   },
+  appearance: false,
 };