+++ src/components/month/Day.vue
@@ -64,7 +64,7 @@
 
   <div
     v-else
-    class="calendar-month__weekday trailing-or-leading"
+    class="space-reserver"
   />
 </template>
 
@@ -101,7 +101,6 @@ export default defineComponent({
       type: Boolean,
       default: false,
     },
-
   },
 
   emits: [
@@ -191,7 +190,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss" scoped>
-.calendar-month__weekday {
+@mixin dayBase {
   height: 100%;
   flex: 1;
   display: flex;
@@ -199,10 +198,15 @@ export default defineComponent({
   align-items: center;
   border-right: var(--qalendar-border-gray-thin);
   border-bottom: var(--qalendar-border-gray-thin);
-  overflow: hidden;
-  transition: background-color 0.2s ease-in-out;
+}
 
-  &.trailing-or-leading {
+.calendar-month__weekday {
+    @include dayBase;
+
+    overflow: hidden;
+    transition: background-color 0.2s ease-in-out;
+
+    &.trailing-or-leading {
     border-right-color: transparent;
 
     + .calendar-month__weekday:not(.trailing-or-leading) {
@@ -222,18 +226,6 @@ export default defineComponent({
     height: auto;
     min-height: 7rem;
     border-right: 0;
-  }
-
-  .calendar-month__week:first-child & {
-    border-top: var(--qalendar-border-gray-thin);
-
-    .qalendar-is-small & {
-      border-top: 0;
-
-      &:first-child {
-        border-top: var(--qalendar-border-gray-thin);
-      }
-    }
   }
 
   .calendar-month__day-name,
@@ -256,6 +248,40 @@ export default defineComponent({
     padding-left: 4px;
     color: var(--qalendar-gray-quite-dark);
     cursor: pointer;
+  }
+}
+
+.space-reserver {
+  @include dayBase;
+
+  border-right-color: transparent;
+
+  + .calendar-month__weekday:not(.trailing-or-leading) {
+    border-left: var(--qalendar-border-gray-thin);
+  }
+}
+
+.space-reserver,
+.trailing-or-leading {
+
+  .qalendar-is-small & {
+    display: none;
+  }
+}
+
+.calendar-month__week:first-child {
+
+  .space-reserver,
+  .calendar-month__weekday {
+    border-top: var(--qalendar-border-gray-thin);
+
+    .qalendar-is-small & {
+      border-top: 0;
+
+      &:first-child {
+        border-top: var(--qalendar-border-gray-thin);
+      }
+    }
   }
 }
 </style>