+++ src/components/week/Day.vue
@@ -16,9 +16,13 @@
       @event-was-dragged="$emit('event-was-dragged', $event)"
       @event-was-resized="handleEventWasResized"
       @drag-start="$emit('drag-start')"
+      @drag-end="$emit('drag-end')"
     >
       <template #weekDayEvent="p">
-        <slot :event-data="p.eventData" name="weekDayEvent"></slot>
+        <slot
+          :event-data="p.eventData"
+          name="weekDayEvent"
+        />
       </template>
     </DayEvent>
 
@@ -91,6 +95,7 @@ export default defineComponent({
     'interval-was-clicked',
     'day-was-clicked',
     'drag-start',
+    'drag-end',
   ],
 
   data() {