+++ src/language/keys.ts
@@ -10,6 +10,8 @@ export const languageKeys = {
     "fr-FR": "Semaine",
     "th-TH": "สัปดาห์",
     "nl-NL": "Week",
+    "ru-RU": "Неделя",
+    "ar-YE": "إسبوع",
   },
   month: {
     "it-IT": "Mese",
@@ -21,6 +23,8 @@ export const languageKeys = {
     "fr-FR": "Mois",
     "th-TH": "เดือน",
     "nl-NL": "Maand",
+    "ru-RU": "Месяц",
+    "ar-YE": "شهر",
   },
   day: {
     "it-IT": "Giorno",
@@ -32,6 +36,8 @@ export const languageKeys = {
     "fr-FR": "Jour",
     "th-TH": "วัน",
     "nl-NL": "Dag",
+    "ru-RU": "День",
+    "ar-YE": "يوم",
   },
 
   /** Other keys */
@@ -45,5 +51,7 @@ export const languageKeys = {
     "fr-FR": "+ d'autres événements",
     "th-TH": "+ เหตุการณ์เพิ่มเติม",
     "nl-NL": "meer evenementen",
+    "ru-RU": "+ ещё события",
+    "ar-YE": "+ المزيد من الأحداث",
   },
 };